pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sys-observe"
version = "0.1.0"
description = "A CLI tool to observe and explain your system (CPU, RAM, Disk, OS)."
authors = [
    { name = "Reviandi Naufal Kurniawan" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "psutil",
    "py-cpuinfo",
    "dataclasses"
]

[project.scripts]
sys-observe = "emc.cli:main"
